move Batch, Job to their own file
batch.py => main.py
move Batch, Job to batch.py
